% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoordinateDescent}
\alias{CoordinateDescent}
\title{Coordinate Descent

Description}
\usage{
CoordinateDescent(X, Y, Beta_0, Lambda, convergence_criteria,
  use_screening_rules = FALSE, use_single_precision = FALSE)
}
\arguments{
\item{X:}{An n x p design matrix.}

\item{Y:}{A 1 x p matrix representing the predictors}

\item{Beta_0:}{A 1 x n matrix that describes the initial guess for Beta}

\item{Lambda:}{The regularization hyper-parameter, ie Lambda*|| Beta ||_1}

\item{convergence_criteria:}{Can be either an positive integer descrbing the number
if iterations that the solve should be run for, or a positive float describing
the smallest allowable Duality Gap.}

\item{use_screening_rules:}{Enable or disable GAPSAFE screening rules. Enabling
these rules MAY speed up the algorithm.}

\item{use_single_precision:}{If set to TRUE double-precision floating point values
will be cast to single-precision. This will result in less memory use and faster
execution, but may result in numerical precision issues.}
}
\value{
A vector containing the coefficients of Beta after the specified convergence
   criteria has been meet.
}
\description{
Use coordinate descent to iteratively solve the LASSO.
}
\details{
In practice Coordinate Descent is often faster than sub-gradient
methods such as ISTA and FISTA.
}
\examples{
# Iterative solving

library(HDIM)

dataset <- matrix(rexp(200, rate=.1), ncol=20)

Y <- dataset[, 1, drop = FALSE]
X <- dataset
Beta_0 <- as.matrix(numeric(ncol(X)))

HDIM::CoordinateDescent( X, Y, Beta_0, 50, as.integer(5) )

# Duality gap convergence criteria.

library(HDIM)

dataset <- matrix(rexp(200, rate=.1), ncol=20)

Y <- dataset[, 1, drop = FALSE]
X <- dataset
Beta_0 <- as.matrix(numeric(ncol(X)))

HDIM::CoordinateDescent( X, Y, Beta_0, 50, 0.1 )
}
\references{
Geoff Gordon & Ryan Tibshirani
  \emph{Coordinate descent},
     \url{https://www.cs.cmu.edu/~ggordon/10725-F12/slides/25-coord-desc.pdf}\cr
}
\seealso{
\code{ISTA} and \code{FISTA}
}
\author{
Benjamin J Phillips e-mail:bejphil@uw.edu
}
